archive/legacy_implementations_20250726/production/requirements.txt
# Editorial Scripts Dependencies
# Generated: January 25, 2025

# Core Web Automation
selenium>=4.15.0
webdriver-manager>=4.0.0

# Data Processing
pandas>=1.5.0
numpy>=1.24.0
requests>=2.28.0
beautifulsoup4>=4.11.0
lxml>=4.9.0

# Configuration and Environment
python-dotenv>=1.0.0
pydantic>=2.0.0

# Academic Data Integration
orcid-python>=0.7.0

# Email and Communication
google-api-python-client>=2.0.0
google-auth-httplib2>=0.1.0
google-auth-oauthlib>=1.0.0

# File Processing
PyPDF2>=3.0.0
python-docx>=0.8.11
openpyxl>=3.1.0

# Date/Time Handling
python-dateutil>=2.8.0
pytz>=2023.3

# HTTP and Networking
urllib3>=1.26.0
certifi>=2023.0.0

# JSON and Data Serialization
jsonschema>=4.17.0

# Testing (Development)
pytest>=7.4.0
pytest-cov>=4.1.0
pytest-mock>=3.11.0

# Browser Drivers (Auto-managed by webdriver-manager)
# Chrome: Automatically downloaded
# Firefox: Automatically downloaded

# macOS Specific (for credential management)
# keyring>=24.0.0  # Only required on macOS

# Optional: Development Tools
black>=23.0.0
flake8>=6.0.0
mypy>=1.5.0

# Version Compatibility
# Python >= 3.8
# macOS >= 10.15 (for Keychain integration)
# Chrome >= 110 (for Selenium compatibility)
# Firefox >= 100 (for Selenium compatibility)